Split out notifiers
- splits out notifiers and writes status for async handler
